+++ ios/Pods/Headers/Private/ABI38_0_0EXVideoThumbnails/ABI38_0_0EXVideoThumbnailsModule.h
@@ -1,1 +1,0 @@
-../../../../versioned-react-native/ABI38_0_0/Expo/EXVideoThumbnails/ABI38_0_0EXVideoThumbnails/ABI38_0_0EXVideoThumbnailsModule.h