+++ ios/Pods/Headers/Private/ABI39_0_0React-Core/ABI39_0_0React/ABI39_0_0RCTFollyConvert.h
@@ -1,1 +1,0 @@
-../../../../../versioned-react-native/ABI39_0_0/ReactNative/React/CxxUtils/ABI39_0_0RCTFollyConvert.h