+++ ios/Pods/Headers/Public/ABI38_0_0React-Core/ABI38_0_0React/ABI38_0_0RCTTextViewManager.h
@@ -1,1 +1,0 @@
-../../../../../versioned-react-native/ABI38_0_0/ReactNative/Libraries/Text/Text/ABI38_0_0RCTTextViewManager.h