+++ ios/Pods/Headers/Public/ABI40_0_0ExpoKit/ABI40_0_0RNSVGPropHelper.h
@@ -1,1 +1,0 @@
-../../../../versioned-react-native/ABI40_0_0/Expo/ExpoKit/Core/Api/Components/Svg/Text/ABI40_0_0RNSVGPropHelper.h