+++ ios/Pods/Headers/Public/EXDevice/EXDevice.h
@@ -1,1 +1,0 @@
-../../../../../packages/expo-device/ios/EXDevice/EXDevice.h