+++ ios/Pods/Target Support Files/ABI39_0_0EXImagePicker/ABI39_0_0EXImagePicker-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_ABI39_0_0EXImagePicker : NSObject
-@end
-@implementation PodsDummy_ABI39_0_0EXImagePicker
-@end