+++ ios/Pods/Target Support Files/EXFacebook/EXFacebook-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_EXFacebook : NSObject
-@end
-@implementation PodsDummy_EXFacebook
-@end