+++ ios/Pods/Target Support Files/GoogleAPIClientForREST/GoogleAPIClientForREST-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_GoogleAPIClientForREST : NSObject
-@end
-@implementation PodsDummy_GoogleAPIClientForREST
-@end