+++ packages/expo-updates/android/src/main/java/expo/modules/updates/launcher/SelectionPolicy.java
@@ -1,11 +1,13 @@
 package expo.modules.updates.launcher;
 
+import org.json.JSONObject;
+
 import expo.modules.updates.db.entity.UpdateEntity;
 
 import java.util.List;
 
 public interface SelectionPolicy {
-  UpdateEntity selectUpdateToLaunch(List<UpdateEntity> updates);
-  List<UpdateEntity> selectUpdatesToDelete(List<UpdateEntity> updates, UpdateEntity launchedUpdate);
-  boolean shouldLoadNewUpdate(UpdateEntity newUpdate, UpdateEntity launchedUpdate);
+  UpdateEntity selectUpdateToLaunch(List<UpdateEntity> updates, JSONObject filters);
+  List<UpdateEntity> selectUpdatesToDelete(List<UpdateEntity> updates, UpdateEntity launchedUpdate, JSONObject filters);
+  boolean shouldLoadNewUpdate(UpdateEntity newUpdate, UpdateEntity launchedUpdate, JSONObject filters);
 }