+++ packages/expo/src/launch/withExpoRoot.tsx
@@ -4,6 +4,18 @@ import * as React from 'react';
 import Notifications from '../Notifications/Notifications';
 import { InitialProps } from './withExpoRoot.types';
 
+// This hook can be optionally imported because __DEV__ never changes during runtime.
+// Using __DEV__ like this enables tree shaking to remove the hook in production.
+let useDevKeepAwake: (tag?: string) => void = () => {};
+
+if (__DEV__) {
+  try {
+    // Optionally import expo-keep-awake
+    const { useKeepAwake } = require('expo-keep-awake');
+    useDevKeepAwake = useKeepAwake;
+  } catch {}
+}
+
 export default function withExpoRoot<P extends InitialProps>(
   AppRootComponent: React.ComponentType<P>
 ): React.ComponentType<P> {
@@ -16,6 +28,8 @@ export default function withExpoRoot<P extends InitialProps>(
 
       didInitialize.current = true;
     }
+
+    useDevKeepAwake();
 
     const combinedProps = {
       ...props,